get rid of filter-box-web and use filter-box component directly
