--- src/components/filter-box-web/index.js
@@ -1,3 +1,0 @@
-import FilterBoxWeb from './filter-box.component'
-
-export default FilterBoxWeb
